third_party/WebKit/LayoutTests/external/wpt/html/webappapis/scripting/processing-model-2/runtime-error-cross-origin-setTimeout-expected.txt

This is a testharness.js-based test.
FAIL window.onerror - runtime error in cross-origin setTimeout assert_equals: first arg expected "Script error." but got "Uncaught ReferenceError: undefined_variable is not defined"
FAIL window.onerror - runtime error in cross-origin setTimeout (column) assert_equals: fourth arg expected 0 but got 1
Harness: the test ran to completion.
